Implement Perspective camera UI

+++ Engine/Resolution.hpp
@@ -3,5 +3,5 @@
 struct Resolution {
 	int w, h;
 
-	inline float Aspect() const { return w / h; }
+	inline float Aspect() const { return static_cast<float>(w) / static_cast<float>(h); }
 };
